Login and sign-up features
I created the login page component on the frontend(html and scss). On the server side in node I created the user model and made 2 post routes for the sign up and login with the parameters of email and password, and 1 get route with the parameters of id and refresh token where we get the access token which is different everytime with each login.