Handling Access and Refresh Token
Finished and tested the handling of the access and refresh tokens. Tested by inspecting the site and manipulating the application tab with all the header parameters. Added the HTTP_INTERCEPTOR to the app.module.ts file and unauthorized users, in the intercept method, that don't have a valid access/refresh token